backend/requirements.txt
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
kaitaistruct==0.10
MarkupSafe==3.0.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.28.1
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
sqlalchemy-orm==1.2.10
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
